refactor: share benchmark primitives
