+++ .gitignore
@@ -66,7 +66,9 @@ logs/
 __pycache__/
 *.py[cod]
 .venv/
+*.egg-info/
 .mypy_cache/
+.pytest_cache/
 .ruff_cache/
 
 # ─────────────────── Claude Code (personal, local) ───────────────────
