+++ experiments/scheduled-task-learning/benchmark-core/benchmark_core/__init__.py
@@ -1,0 +1,1 @@
+"""Scenario-neutral deterministic benchmark primitives."""
