+++ experiments/scheduled-task-learning/page-change/page_benchmark/feedback.py
@@ -5,7 +5,11 @@ from __future__ import annotations
 import argparse
 from typing import Any
 
+from benchmark_core.feedback import normalize_feedback
+
 from .canonical import SHA256_HEX, canonical_sha256, dumps, load_object, write
+
+__all__ = ["normalize_feedback"]
 
 FEEDBACK_GENERATOR_VERSION = "page-feedback-v1"
 
@@ -29,33 +33,6 @@ def feedback_generator_identity(
         "sha256": category_sha256,
         "templates_sha256": canonical_sha256(templates_contract),
     }
-
-
-def normalize_feedback(
-    runs: list[dict[str, Any]],
-    templates_contract: dict[str, Any],
-) -> list[dict[str, Any]]:
-    templates = templates_contract["templates"]
-    normalized: list[dict[str, Any]] = []
-    for run in sorted(runs, key=lambda item: item["run_id"]):
-        run_id = run["run_id"]
-        for index, entry in enumerate(run["score_result"]["error_ledger"]):
-            code = entry["code"]
-            if code not in templates:
-                raise ValueError(f"no frozen feedback template for error code {code}")
-            template = templates[code]
-            item: dict[str, Any] = {
-                "run_id": run_id,
-                "ledger_index": index,
-                "code": code,
-                "critical": entry["critical"],
-                "summary": template["summary"],
-                "guidance": template["guidance"],
-            }
-            if "target_class" in entry:
-                item["target_class"] = entry["target_class"]
-            normalized.append(item)
-    return normalized
 
 
 def main() -> None:
