+++ experiments/scheduled-task-learning/page-change/scripts/lint.sh
@@ -6,7 +6,7 @@
 set -euo pipefail
 cd "$(dirname "$0")/.."
 
-paths=(page_benchmark tests)
+paths=(page_benchmark tests ../benchmark-core/benchmark_core)
 
 if [[ "${1:-}" == "--fix" ]]; then
   uv run ruff format "${paths[@]}"
@@ -18,5 +18,6 @@ fi
 uv run ruff format --check "${paths[@]}"
 uv run ruff check "${paths[@]}"
 uv run mypy
+uv run mypy --config-file ../benchmark-core/pyproject.toml ../benchmark-core/benchmark_core
 
 echo "lint: ok"
